+++ src/screens/transaction_details/components/operations/components/desktop/index.tsx
@@ -1,20 +1,27 @@
 import React from 'react';
+import * as R from 'ramda';
 import useTranslation from 'next-translate/useTranslation';
+import { formatNumber } from '@utils/format_token';
+import Link from 'next/link';
 import {
   Table,
   TableHead,
   TableRow,
   TableCell,
   TableBody,
+  Typography,
 } from '@material-ui/core';
 import { AvatarName } from '@components';
 import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
+import { TOKEN_DETAILS } from '@utils/go_to_page';
 import { columns } from './utils';
 import { OperationType } from '../../../../types';
 
 const Desktop: React.FC<{items: OperationType[]} & ComponentDefault> = (props) => {
   const { t } = useTranslation('transactions');
   const formattedItems = props.items.map((x) => {
+    const isToken = R.pathOr('', ['identifier'], x).split('-').length === 2;
+
     return ({
       action: x.action.replace(/([A-Z])/g, ' $1').toUpperCase(),
       identifier: x.identifier,
@@ -22,7 +29,7 @@ const Desktop: React.FC<{items: OperationType[]} & ComponentDefault> = (props) =
         <AvatarName
           address={x.sender}
           name={getMiddleEllipsis(x.sender, {
-            beginning: 13, ending: 15,
+            beginning: 15, ending: 20,
           })}
         />
       ),
@@ -30,9 +37,26 @@ const Desktop: React.FC<{items: OperationType[]} & ComponentDefault> = (props) =
         <AvatarName
           address={x.receiver}
           name={getMiddleEllipsis(x.receiver, {
-            beginning: 13, ending: 15,
+            beginning: 15, ending: 20,
           })}
         />
+      ),
+      value: (
+        isToken ? (
+          <div>
+            <Typography component="span">
+              {formatNumber(x.value.value, x.value.exponent)}
+              {' '}
+            </Typography>
+            <Link href={TOKEN_DETAILS(x.identifier)} passHref>
+              <Typography component="a">
+                {x.value.displayDenom.toUpperCase()}
+              </Typography>
+            </Link>
+          </div>
+        ) : (
+          `${formatNumber(x.value.value, x.value.exponent)} ${x.value.displayDenom.toUpperCase()}`
+        )
       ),
     });
   });