+++ src/screens/transaction_details/components/operations/components/mobile/index.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
+import * as R from 'ramda';
 import useTranslation from 'next-translate/useTranslation';
-import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
+import { formatNumber } from '@utils/format_token';
+import Link from 'next/link';
+import { TOKEN_DETAILS } from '@utils/go_to_page';
 import {
   Typography, Divider,
 } from '@material-ui/core';
@@ -12,24 +15,42 @@ const Mobile: React.FC<{items: OperationType[]} & ComponentDefault> = (props) =>
   const { t } = useTranslation('transactions');
   const classes = useStyles();
   const formattedItems = props.items.map((x) => {
+    const isToken = R.pathOr('', ['identifier'], x).split('-').length === 2;
     return ({
       action: x.action.replace(/([A-Z])/g, ' $1').toUpperCase(),
       identifier: x.identifier,
       sender: (
         <AvatarName
           address={x.sender}
-          name={getMiddleEllipsis(x.sender, {
-            beginning: 13, ending: 15,
-          })}
+          name={x.sender}
         />
       ),
       receiver: (
         <AvatarName
           address={x.receiver}
-          name={getMiddleEllipsis(x.receiver, {
-            beginning: 13, ending: 15,
-          })}
+          name={x.receiver}
         />
+      ),
+      value: (
+        isToken ? (
+          <div>
+            <Typography component="span">
+              {formatNumber(x.value.value, x.value.exponent)}
+              {' '}
+            </Typography>
+            <Link href={TOKEN_DETAILS(x.identifier)} passHref>
+              <Typography component="a">
+                {x.value.displayDenom.toUpperCase()}
+              </Typography>
+            </Link>
+          </div>
+        ) : (
+          <Typography>
+            {formatNumber(x.value.value, x.value.exponent)}
+            {' '}
+            {x.value.displayDenom.toUpperCase()}
+          </Typography>
+        )
       ),
     });
   });
@@ -57,17 +78,13 @@ const Mobile: React.FC<{items: OperationType[]} & ComponentDefault> = (props) =>
                 <Typography variant="h4" className="label">
                   {t('receiver')}
                 </Typography>
-                <Typography variant="body1" className="value">
-                  {x.receiver}
-                </Typography>
+                {x.receiver}
               </div>
               <div className={classes.item}>
                 <Typography variant="h4" className="label">
-                  {t('identifier')}
+                  {t('value')}
                 </Typography>
-                <Typography variant="body1" className="value">
-                  {x.identifier}
-                </Typography>
+                {x.value}
               </div>
             </div>
             {i !== formattedItems.length - 1 && <Divider />}