+++ src/screens/transaction_details/types.ts
@@ -26,6 +26,7 @@ export type OperationType = {
   sender: string;
   receiver: string;
   identifier: string;
+  value: TokenUnit;
 }
 
 export type ResultType = {